tools/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-magic==0.4.27
sniffio==1.3.1
